+++ src/types.ts
@@ -6,8 +6,9 @@ export interface GlobalInfinityQuery {
   id: string;
   query: InfinityQuery;
 }
-export type AuthType = 'none' | 'basicAuth' | 'oauthPassThru' | 'oauth2';
+export type AuthType = 'none' | 'basicAuth' | 'apiKey' | 'bearerToken' | 'oauthPassThru' | 'oauth2';
 export type OAuth2Type = 'client_credentials' | 'jwt' | 'others';
+export type APIKeyType = 'header' | 'query';
 export type OAuth2Props = {
   oauth2_type?: OAuth2Type;
   client_id?: string;
@@ -19,6 +20,8 @@ export type OAuth2Props = {
 };
 export interface InfinityOptions extends DataSourceJsonData {
   auth_method?: AuthType;
+  apiKeyKey?: string;
+  apiKeyType?: APIKeyType;
   oauth2?: OAuth2Props;
   tlsSkipVerify?: boolean;
   tlsAuth?: boolean;
@@ -27,6 +30,7 @@ export interface InfinityOptions extends DataSourceJsonData {
   global_queries?: GlobalInfinityQuery[];
   timeoutInSeconds?: number;
   oauthPassThru?: boolean;
+  allowedHosts?: string[];
 }
 
 export interface InfinitySecureOptions {
@@ -34,6 +38,8 @@ export interface InfinitySecureOptions {
   tlsCACert?: string;
   tlsClientCert?: string;
   tlsClientKey?: string;
+  apiKeyValue?: string;
+  bearerToken?: string;
   oauth2ClientSecret?: string;
   oauth2JWTPrivateKey?: string;
 }